Simplify idb usage and collapse the class
