+++ packages/objectloader2/src/core/interfaces.ts
@@ -12,7 +12,7 @@ export interface Downloader extends Queue<string> {
 }
 
 export interface Database {
-  getAll(keys: string[]): Promise<(Item | undefined)[]>
-  saveBatch(params: { batch: Item[] }): Promise<void>
-  disposeAsync(): Promise<void>
+  getAll(ids: string[]): Promise<(Item | undefined)[]>
+  putAll(batch: Item[]): Promise<void>
+  dispose(): void
 }
