+++ packages/objectloader2/src/core/objectLoader2Factory.ts
@@ -1,7 +1,7 @@
 import { CustomLogger, getFeatureFlag, ObjectLoader2Flags } from '../types/functions.js'
 import { Base } from '../types/types.js'
 import { ObjectLoader2 } from './objectLoader2.js'
-import IndexedDatabase from './stages/indexedDatabase.js'
+import { IndexedDatabase } from './stages/indexedDatabase.js'
 import { MemoryDatabase } from './stages/memory/memoryDatabase.js'
 import { MemoryDownloader } from './stages/memory/memoryDownloader.js'
 import ServerDownloader from './stages/serverDownloader.js'
